+++ hahomematic/client.py
@@ -390,11 +390,11 @@ class Client(ABC):
     ) -> bool:
         """Set single value on paramset VALUES."""
         try:
+            _LOGGER.debug("_set_value: %s, %s, %s", channel_address, parameter, value)
             if rx_mode:
                 await self._proxy.setValue(channel_address, parameter, value, rx_mode)
             else:
                 await self._proxy.setValue(channel_address, parameter, value)
-            _LOGGER.debug("_set_value: %s, %s, %s", channel_address, parameter, value)
         except BaseHomematicException as hhe:
             _LOGGER.warning(
                 "_set_value failed with %s [%s]: %s, %s, %s",
@@ -415,7 +415,7 @@ class Client(ABC):
         value: Any,
         rx_mode: str | None = None,
     ) -> bool:
-        """Set single value on paramset VALUES."""
+        """Set single value on paramset VALUES. #CC"""
         if paramset_key == PARAMSET_KEY_VALUES:
             return await self._set_value(
                 channel_address=channel_address,
@@ -463,7 +463,7 @@ class Client(ABC):
         """
         Set paramsets manually.
         Address is usually the channel_address,
-        but for bidcos devices there is a master paramset at the device.
+        but for bidcos devices there is a master paramset at the device. #CC
         """
         try:
             _LOGGER.debug("put_paramset: %s, %s, %s", address, paramset_key, value)
@@ -845,7 +845,7 @@ class ClientHomegear(Client):
         return "Homegear_SN0815"
 
 
-class ClientLocal(Client):
+class ClientLocal(Client):  # pragma: no cover
     """
     Local client object that emulates the XML-RPC proxy
     and provides access to other data via XML-RPC