+++ changelog.md
@@ -3,6 +3,8 @@
 - Remove support for python 3.9
 - Remove old-style union syntax
 - Remove to int converter for HmIP-SCTH230 CONCENTRATION
+- Remove set_value, put_paramset from central
+- Remove put_paramset from custom_entity
 
 # Version 2023.1.7 (2023-01-24)
 - Aggregate calls to backend